add timestamp to puts
The problem I am facing here is that when 'snapshots' are taken of a wallet and stored in balanceHistoryBtc they are usually wrong because another worker ahead of it has already modified the db. This does not matter to inputs and wallets since the issue is usually resolved in way less than a second when the worker catches up. For example if a balance is negative because the send out is recorded before the transfer in. the slow earlier worker will still add the transfer in to the total balance creating the correct value. At the top of the chain (99.9% of the time) this won't be an issue. If I add a timestamp to input/output I can get the balance at any given time with a LONG query. Can can probably write a script to fill new entity once at top of chain. Hi future me